Fix waiting for apps in click_next_page() method
