+++ pages/desktop/developer_hub/developer_submissions.py
@@ -89,7 +89,6 @@ class DeveloperSubmissions(Base):
 
     @property
     def paginator(self):
-        from pages.desktop.regions.paginator import Paginator
         return Paginator(self.testsetup)
 
 
@@ -181,3 +180,73 @@ class Sorter(Page):
             for option in self._sorter.find_elements(*self._options_locator):
                 if option.text.lower() == value.lower():
                     option.click()
+
+class Paginator(Page):
+
+    _paginator_locator = (By.CSS_SELECTOR, 'nav.paginator')
+    _apps_locator = (By.CSS_SELECTOR, 'div.items > div.item')
+
+    #Numbering
+    _page_number_locator = (By.CSS_SELECTOR, 'nav.paginator .num > a:nth-child(1)')
+    _total_page_number_locator = (By.CSS_SELECTOR, 'nav.paginator .num > a:nth-child(2)')
+
+    #Navigation
+    _first_page_locator = (By.CSS_SELECTOR, 'nav.paginator .rel a:nth-child(1)')
+    _prev_locator = (By.CSS_SELECTOR, 'nav.paginator .rel a.prev')
+    _next_locator = (By.CSS_SELECTOR, 'nav.paginator .rel a.next')
+    _last_page_locator = (By.CSS_SELECTOR, 'nav.paginator .rel a:nth-child(4)')
+
+    #Position
+    _start_item_number_locator = (By.CSS_SELECTOR, 'nav.paginator .pos b:nth-child(1)')
+    _end_item_number_locator = (By.CSS_SELECTOR, 'nav.paginator .pos b:nth-child(2)')
+    _total_item_number = (By.CSS_SELECTOR, 'nav.paginator .pos b:nth-child(3)')
+
+    def wait_for_apps_visible(self):
+        self.wait_for_element_visible(*self._apps_locator)
+
+    @property
+    def is_paginator_present(self):
+        return self.is_element_present(*self._paginator_locator)
+
+    @property
+    def page_number(self):
+        return int(self.selenium.find_element(*self._page_number_locator).text)
+
+    @property
+    def total_page_number(self):
+        if self.is_paginator_present:
+            return int(self.selenium.find_element(*self._total_page_number_locator).text)
+        else:
+            return 1
+
+    @property
+    def is_prev_page_disabled(self):
+        return 'disabled' in self.selenium.find_element(*self._prev_locator).get_attribute('class')
+
+    @property
+    def is_first_page_disabled(self):
+        return 'disabled' in self.selenium.find_element(*self._first_page_locator).get_attribute('class')
+
+    def click_next_page(self):
+        self.selenium.find_element(*self._next_locator).click()
+        self.wait_for_apps_visible()
+
+    @property
+    def is_next_page_disabled(self):
+        return 'disabled' in self.selenium.find_element(*self._next_locator).get_attribute('class')
+
+    @property
+    def is_last_page_disabled(self):
+        return 'disabled' in self.selenium.find_element(*self._last_page_locator).get_attribute('class')
+
+    @property
+    def start_item(self):
+        return int(self.selenium.find_element(*self._start_item_number_locator).text)
+
+    @property
+    def end_item(self):
+        return int(self.selenium.find_element(*self._end_item_number_locator).text)
+
+    @property
+    def total_items(self):
+        return int(self.selenium.find_element(*self._total_item_number).text)
